# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST LEGALTECH - TEMIS",
        "description": "Mostando la Lista de URI's de mi API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
...
        "/api/notificacion-reportes": {
            "get": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Obtener notificaciones de reportes",
                "operationId": "e7280861acbfa3ef71c8906242b6fb40",
                "responses": {
                    "200": {
                        "description": "Lista de notificaciones de reportes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notificaciones"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_history": {
                                                        "type": "integer",
                                                        "example": 16
                                                    },
                                                    "id_exp": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_client": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_movimiento": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "date_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-17T00:15:10.000000Z"
                                                    },
                                                    "estado": {
                                                        "type": "string",
                                                        "example": "no"
                                                    },
                                                    "entidad": {
                                                        "type": "string",
                                                        "example": "Sinoe"
                                                    },
                                                    "exp_n_exp": {
                                                        "type": "string",
                                                        "example": "10600-2023-0-1706-JR-PE-01"
                                                    },
                                                    "movi_accion_realizada": {
                                                        "type": "string",
                                                        "example": "RES. N° UNO MAS ESCRITO DE REQUERIMIENTO DE ACUSACION DIRECTA Y ANEXOS"
                                                    },
                                                    "movi_anotaciones": {
                                                        "type": "string",
                                                        "example": "1 JUZGADO DE INVEST. PREPARATORIA-MBJ JLO"
                                                    },
                                                    "client_type_contact": {
                                                        "type": "string",
                                                        "example": "Empresa"
                                                    },
                                                    "client_name": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "client_last_name": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "client_name_company": {
                                                        "type": "string",
                                                        "example": "demo"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "/seguimientos-sinoe?Exp=1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-estado-history-movements/{id}": {
            "put": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Actualizar estado de notificación de reporte por ID",
                "operationId": "ebb3ccc3391ffcfefac28d34716bf0cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la notificación",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de la notificación actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "id": {
                                            "type": "string",
                                            "example": "16"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El estado de la notificación ha sido actualizado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recurso no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación Bearer",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}